Support Java 11 JDK: replace maven plugin and legacy runtime libs
Cobertura is a maven plugin for test reporting. Cobertura is going to be
abandoned in terms of supporting Java versions >= 8. In order to natively
support Java 11 a different maven plugin was already used in the Jenkinsfile,
Jacoco. Jacoco is supports Java 11 out of the box. Only the already used
version needed a small upgrade to work.

Additionally, some runtime libraries are no longer bundled with Java 11
(like com.sun.* or Java EE libraries). These dependencies are now added
as explicit dependencies in the pom file.